fssb: Complete overhaul of the memory write thingy
Phew. Ever since I wrote it the first time, I had this nagging
feeling at the back of my head -- never overwrite user memory!
Till now, there wasn't a problem - open(2) and unlink(2) are
functions that take in just a single string parameter. But later
on, I'll implement support for rename(2), which takes in two
different strings.

When I overwrite both of them, it *may* overlap and overwrite
each other; which obviously is a no-no.

Anyway, this is a much safer approach :)